+++ changelog.md
@@ -21,6 +21,7 @@
 - Seperate K3S Server and Agent Memory and CPU resource definition.
 - Change K3S to use docker engine
 - Add DOCKER_SHM_SIZE variable to configure shm-size.
+- Increase etcd election timeout
 
 ## v0.2.0